refactor CLI classes into individual files, create Shell directory
