+++ framework/classes.php
@@ -222,6 +222,16 @@ return [
 'TUser' => 'Prado\Security\TUser',
 'TUserManager' => 'Prado\Security\TUserManager',
 'TUserManagerPasswordMode' => 'Prado\Security\TUserManagerPasswordMode',
+'TActiveRecordGenAction' => 'Prado\Shell\Actions\TActiveRecordGenAction',
+'TActiveRecordGenAllAction' => 'Prado\Shell\Actions\TActiveRecordGenAllAction',
+'TAppAction' => 'Prado\Shell\Actions\TAppAction',
+'TAppHelpAction' => 'Prado\Shell\Actions\TAppHelpAction',
+'TFlushCachesAction' => 'Prado\Shell\Actions\TFlushCachesAction',
+'TPphShellAction' => 'Prado\Shell\Actions\TPphShellAction',
+'TShellAction' => 'Prado\Shell\TShellAction',
+'TShellAppAction' => 'Prado\Shell\TShellAppAction',
+'TShellApplication' => 'Prado\Shell\TShellApplication',
+'TShellInterpreter' => 'Prado\Shell\TShellInterpreter',
 'TApplication' => 'Prado\TApplication',
 'TApplicationComponent' => 'Prado\TApplicationComponent',
 'TApplicationConfiguration' => 'Prado\TApplicationConfiguration',
@@ -235,7 +245,6 @@ return [
 'TModule' => 'Prado\TModule',
 'TPropertyValue' => 'Prado\TPropertyValue',
 'TService' => 'Prado\TService',
-'TShellApplication' => 'Prado\TShellApplication',
 'TBehaviorParameterLoader' => 'Prado\Util\Behaviors\TBehaviorParameterLoader',
 'IBaseBehavior' => 'Prado\Util\IBaseBehavior',
 'IBehavior' => 'Prado\Util\IBehavior',
